# Optimize multiple hkdf reads case